feat(interface): initial version -> v1 folder
