+++ src/vm.rs
@@ -536,7 +536,7 @@ fn write_boot_info_to_mem(
 				)
 			)]
 			serial_port_base: SerialPortBase::new(
-				(uhyve_interface::HypercallAddress::Uart as u16).into(),
+				(uhyve_interface::v1::HypercallAddress::Uart as u16).into(),
 			),
 			device_tree: Some((mem.guest_addr().as_u64() + FDT_OFFSET).try_into().unwrap()),
 		},
